requirements/requirements.txt
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
google-api-core==2.0.1
google-api-python-client==2.23.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-cloud-pubsub==2.8.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.0
httplib2==0.19.1
idna==3.2
libcst==0.3.21
mypy-extensions==0.4.3
packaging==21.0
proto-plus==1.19.1
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymsteams==0.1.15
pyparsing==2.4.7
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
typing-extensions==3.10.0.2
typing-inspect==0.7.1
uritemplate==3.0.1
urllib3==1.26.7
